# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PheLa86 Game Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  username: string;
  balance: number;
  goldCoins: number;
  level: number;
  vipLevel: number;
  /** @nullable */
  avatar: string | null;
  wins: number;
  losses: number;
  winRate: number;
  joinedAt: string;
}

export interface PlayerUpdate {
  username?: string;
  avatar?: string;
}

export type RoomGameType = typeof RoomGameType[keyof typeof RoomGameType];


export const RoomGameType = {
  phela: 'phela',
  taixiu: 'taixiu',
  baucua: 'baucua',
  poker: 'poker',
  tienlen: 'tienlen',
} as const;

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  waiting: 'waiting',
  playing: 'playing',
  full: 'full',
} as const;

export interface Room {
  id: number;
  name: string;
  gameType: RoomGameType;
  minBet: number;
  maxBet: number;
  playerCount: number;
  maxPlayers: number;
  status: RoomStatus;
  isHot: boolean;
  isVip: boolean;
  hasPassword: boolean;
  /** @nullable */
  jackpot: number | null;
}

export type RoomInputGameType = typeof RoomInputGameType[keyof typeof RoomInputGameType];


export const RoomInputGameType = {
  phela: 'phela',
  taixiu: 'taixiu',
  baucua: 'baucua',
  poker: 'poker',
  tienlen: 'tienlen',
} as const;

export interface RoomInput {
  name: string;
  gameType: RoomInputGameType;
  minBet: number;
  maxBet: number;
  /** @nullable */
  password?: string | null;
  maxPlayers?: number;
}

export interface LeaderboardEntry {
  rank: number;
  playerId: number;
  username: string;
  /** @nullable */
  avatar: string | null;
  totalWinnings: number;
  gamesPlayed: number;
  winRate: number;
  level: number;
}

export interface LeaderboardSummary {
  totalPlayers: number;
  totalGamesPlayed: number;
  biggestWin: number;
  topPlayer: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  win: 'win',
  loss: 'loss',
  bonus: 'bonus',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  status: TransactionStatus;
  createdAt: string;
  description: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  amount: number;
}

export interface GameType {
  id: string;
  name: string;
  displayName: string;
  description: string;
  minBet: number;
  maxBet: number;
  activeRooms: number;
  icon: string;
}

export interface LobbyStats {
  onlinePlayers: number;
  activeRooms: number;
  jackpotTotal: number;
  totalGamesToday: number;
}

export type ListRoomsParams = {
gameType?: ListRoomsGameType;
};

export type ListRoomsGameType = typeof ListRoomsGameType[keyof typeof ListRoomsGameType];


export const ListRoomsGameType = {
  phela: 'phela',
  taixiu: 'taixiu',
  baucua: 'baucua',
  poker: 'poker',
  tienlen: 'tienlen',
  all: 'all',
} as const;

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  alltime: 'alltime',
} as const;