# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PheLa86 Game Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all game rooms
      parameters:
        - name: gameType
          in: query
          schema:
            type: string
            enum: [phela, taixiu, baucua, poker, tienlen, all]
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Created room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, name, gameType, minBet, maxBet, playerCount, maxPlayers, status, isHot, isVip, hasPassword, jackpot]
      properties:
        id:
          type: integer
        name:
          type: string
        gameType:
          type: string
          enum: [phela, taixiu, baucua, poker, tienlen]
        minBet:
          type: number
        maxBet:
          type: number
        playerCount:
          type: integer
        maxPlayers:
          type: integer
        status:
          type: string
          enum: [waiting, playing, full]
        isHot:
          type: boolean
        isVip:
          type: boolean
        hasPassword:
          type: boolean
        jackpot:
          type: ["number", "null"]
    RoomInput:
      type: object
      required: [name, gameType, minBet, maxBet]
      properties:
        name:
          type: string
        gameType:
          type: string
          enum: [phela, taixiu, baucua, poker, tienlen]
        minBet:
          type: number
        maxBet:
          type: number
        password:
          type: ["string", "null"]
        maxPlayers:
          type: integer
...