# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PheLa86 Game Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomGameType } from './roomGameType';
import type { RoomStatus } from './roomStatus';

export interface Room {
  id: number;
  name: string;
  gameType: RoomGameType;
  minBet: number;
  maxBet: number;
  playerCount: number;
  maxPlayers: number;
  status: RoomStatus;
  isHot: boolean;
  isVip: boolean;
  hasPassword: boolean;
  /** @nullable */
  jackpot: number | null;
}